test: isolate Fabric auth retry coverage
Move auth provenance coverage into a focused test file so the master patch replays cleanly onto the spark4.1 release branch.

AB#3582121


Copilot-Session: 98752775-8e28-43eb-9609-2249dd586f80
